Get rid of archiver exceptions and handle errors explicity
